backend/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.11.0
bcrypt==5.0.0
beautifulsoup4==4.14.2
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
curl_cffi==0.13.0
ecdsa==0.19.1
fastapi==0.121.1
frozendict==2.4.6
greenlet==3.2.4
h11==0.16.0
idna==3.11
multitasking==0.0.12
numpy==2.3.4
pandas==2.3.3
peewee==3.18.3
platformdirs==4.5.0
protobuf==6.33.0
pyasn1==0.6.1
pycparser==2.23
pydantic==2.12.4
pydantic_core==2.41.5
python-dateutil==2.9.0.post0
python-jose==3.5.0
pytz==2025.2
requests==2.32.5
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
soupsieve==2.8
SQLAlchemy==2.0.44
starlette==0.49.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.38.0
websockets==15.0.1
yfinance==0.2.66

python-multipart
